Move implementation to separate files
