+++ lib/appconfig/configuration_profile.ts
@@ -1,0 +1,43 @@
+import * as cdk from '@aws-cdk/core';
+import * as appconfig from '@aws-cdk/aws-appconfig';
+
+import { Application } from './application';
+
+export interface IConfigurationProfile extends cdk.IResource {
+  readonly configurationProfileId: string;
+}
+
+export interface ConfigurationProfileProps {
+  readonly application: Application;
+  readonly name: string;
+  readonly description?: string;
+}
+
+export class HostedConfigurationProfile extends cdk.Resource implements IConfigurationProfile, cdk.ITaggable {
+  public readonly configurationProfileId: string;
+  public readonly tags: cdk.TagManager;
+  private readonly resource: appconfig.CfnConfigurationProfile;
+
+  constructor(scope: cdk.Construct, id: string, props: ConfigurationProfileProps) {
+    super(scope, id, {
+      physicalName: props.name
+    });
+
+    this.tags = new cdk.TagManager(cdk.TagType.STANDARD, 'AWS::AppConfig::ConfigurationProfile');
+
+    this.resource = new appconfig.CfnConfigurationProfile(this, 'Resource', {
+      applicationId: props.application.applicationId,
+      name: props.name,
+      description: props.description,
+      locationUri: 'hosted'
+      // todo: validators
+      // validators: [],
+    });
+
+    this.configurationProfileId = this.resource.ref;
+  }
+
+  protected prepare() {
+    this.resource.tags = this.tags.renderTags();
+  }
+}
